updated folder structure, and updated readme
