+++ aerial_imagery/.idea/workspace.xml
@@ -2,13 +2,10 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="89d60d08-afca-4c69-b2c2-e25c03e91d02" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/.idea/other.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/aerial_imagery.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/aerial_imagery.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/overhead_aerial_imagery.py" beforeDir="false" afterPath="$PROJECT_DIR$/overhead_aerial_imagery.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../keras/.idea/keras.iml" beforeDir="false" afterPath="$PROJECT_DIR$/../keras/.idea/keras.iml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../keras/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../keras/.idea/misc.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../keras/.idea/vcs.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../keras/.idea/vcs.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../keras/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../keras/.idea/workspace.xml" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
     <option name="SHOW_DIALOG" value="false" />
@@ -35,8 +32,8 @@
       <file pinned="false" current-in-tab="true">
         <entry file="file://$PROJECT_DIR$/overhead_aerial_imagery.py">
           <provider selected="true" editor-type-id="text-editor">
-            <state relative-caret-position="347">
-              <caret line="104" column="34" selection-start-line="104" selection-start-column="34" selection-end-line="104" selection-end-column="34" />
+            <state relative-caret-position="101">
+              <caret line="105" column="42" selection-start-line="105" selection-start-column="42" selection-end-line="105" selection-end-column="42" />
               <folding>
                 <element signature="e#653#662#0" expanded="true" />
               </folding>
@@ -56,7 +53,7 @@
       </list>
     </option>
   </component>
-  <component name="ProjectFrameBounds" extendedState="4">
+  <component name="ProjectFrameBounds" extendedState="6">
     <option name="y" value="17" />
     <option name="width" value="960" />
     <option name="height" value="997" />
@@ -87,7 +84,7 @@
     <property name="last_opened_file_path" value="$PROJECT_DIR$/../../trader" />
     <property name="nodejs_interpreter_path.stuck_in_default_project" value="undefined stuck path" />
     <property name="nodejs_npm_path_reset_for_default_project" value="true" />
-    <property name="settings.editor.selected.configurable" value="reference.settingsdialog.IDE.editor.colors" />
+    <property name="settings.editor.selected.configurable" value="com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable" />
   </component>
   <component name="RunDashboard">
     <option name="ruleStates">
@@ -144,11 +141,12 @@
       <workItem from="1554628436870" duration="17191000" />
       <workItem from="1554790756682" duration="1364000" />
       <workItem from="1554792137191" duration="1234000" />
+      <workItem from="1554882965467" duration="904000" />
     </task>
     <servers />
   </component>
   <component name="TimeTrackingManager">
-    <option name="totallyTimeSpent" value="19789000" />
+    <option name="totallyTimeSpent" value="21214000" />
   </component>
   <component name="TodoView">
     <todo-panel id="selected-file">
@@ -160,7 +158,7 @@
     </todo-panel>
   </component>
   <component name="ToolWindowManager">
-    <frame x="0" y="17" width="960" height="997" extended-state="4" />
+    <frame x="0" y="23" width="1280" height="720" extended-state="6" />
     <editor active="true" />
     <layout>
       <window_info content_ui="combo" id="Project" order="0" weight="0.12540022" />
@@ -172,7 +170,7 @@
       <window_info anchor="bottom" id="Debug" order="3" weight="0.4" />
       <window_info anchor="bottom" id="Cvs" order="4" weight="0.25" />
       <window_info anchor="bottom" id="Inspection" order="5" weight="0.4" />
-      <window_info anchor="bottom" id="Python Console" order="6" visible="true" weight="0.3267442" />
+      <window_info anchor="bottom" id="Python Console" order="6" visible="true" weight="0.32643312" />
       <window_info anchor="bottom" id="TODO" order="7" weight="0.32906976" />
       <window_info anchor="bottom" id="Docker" order="8" show_stripe_button="false" />
       <window_info anchor="bottom" id="Database Changes" order="9" show_stripe_button="false" />
@@ -192,8 +190,8 @@
   <component name="editorHistoryManager">
     <entry file="file://$PROJECT_DIR$/overhead_aerial_imagery.py">
       <provider selected="true" editor-type-id="text-editor">
-        <state relative-caret-position="347">
-          <caret line="104" column="34" selection-start-line="104" selection-start-column="34" selection-end-line="104" selection-end-column="34" />
+        <state relative-caret-position="101">
+          <caret line="105" column="42" selection-start-line="105" selection-start-column="42" selection-end-line="105" selection-end-column="42" />
           <folding>
             <element signature="e#653#662#0" expanded="true" />
           </folding>
