--- keras/data.csv
@@ -1,31 +1,0 @@
-date_elected,party,prev_held_office,previous_party_1,previous_party_2,was_vp_or_vp_runner
-1900-11-06,1,1,1,0,0
-1904-11-08,1,0,1,1,1
-1908-11-03,1,0,1,1,0
-1912-11-05,0,0,1,1,0
-1916-11-07,0,1,0,1,0
-1920-11-02,1,0,0,0,0
-1924-11-04,1,0,1,0,1
-1928-11-06,1,0,1,1,0
-1932-11-08,0,0,1,1,1
-1936-11-03,0,1,0,1,1
-1940-11-05,0,1,0,0,1
-1944-11-07,0,1,0,0,1
-1948-11-02,0,0,0,0,1
-1952-11-04,1,0,0,0,0
-1956-11-06,1,1,1,0,0
-1960-11-08,0,0,1,1,0
-1964-11-03,0,0,0,1,1
-1968-11-05,1,0,0,0,1
-1972-11-07,1,1,1,0,1
-1976-11-02,0,0,1,1,0
-1980-11-04,1,0,0,1,0
-1984-11-06,1,1,1,0,0
-1988-11-08,1,0,1,1,1
-1992-11-03,0,0,1,1,0
-1996-11-05,0,1,0,1,0
-2000-11-07,1,0,0,0,0
-2004-11-02,1,1,1,0,0
-2008-11-04,0,0,1,1,0
-2012-11-06,0,1,0,1,0
-2016-11-08,1,0,0,0,0
